backend/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "briefbiz-backend"
version = "0.1.0"
description = "Backend services for BriefBiz"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "aiosqlite>=0.20.0",
  "alembic>=1.13.2",
  "asyncpg>=0.29.0",
  "bcrypt>=4.2.0",
  "celery[redis]>=5.4.0",
  "email-validator>=2.2.0",
  "elasticsearch[async]>=8.13.0",
  "feedparser>=6.0.11",
  "fastapi>=0.115.0",
  "greenlet>=3.0.3",
  "httpx>=0.27.0",
  "jinja2>=3.1.4",
  "openai>=1.54.0",
  "passlib[bcrypt]>=1.7.4",
  "pillow>=11.1.0",
  "pydantic-settings>=2.4.0",
  "PyJWT>=2.9.0",
  "python-dateutil>=2.9.0.post0",
  "python-dotenv>=1.0.1",
  "redis>=5.0.8",
  "sqlalchemy>=2.0.32",
  "structlog>=24.4.0",
  "uvicorn[standard]>=0.30.6",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.2",
  "pytest-asyncio>=0.23.8",
  "pytest-mock>=3.14.0",
  "ruff>=0.6.2",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py311"
